requirements.txt
Flask
whisper
openai
sqlalchemy
psycopg2-binary
python-dotenv
